+++ app/Services/Role/RoleService.php
@@ -5,23 +5,52 @@ namespace App\Services\Role;
 
 use App\Repositories\Role\RoleRepositoryInterface;
 use App\Services\Contract\MyService;
+use App\Validations\RoleValidation;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 
 class RoleService extends MyService
 {
     private $roleRepo;
+    private $roleValidation;
 
-    public function __construct(RoleRepositoryInterface $roleRepo)
+    public function __construct(RoleRepositoryInterface $roleRepo, RoleValidation $roleValidation)
     {
         $this->roleRepo = $roleRepo;
+        $this->roleValidation = $roleValidation;
     }
 
     public function index()
     {
         $list = $this->getAll();
-        $listFeature = $this->getAllFeature();
+        $listApiName = $this->getAllFeatureApiName();
 
-        return view('/role/list')->with(['list' => $list, 'listFeature' => $listFeature]);
+        return view('/role/list')->with(['list' => $list, 'listApiName' => $listApiName]);
+    }
+
+    public function detail($id)
+    {
+        $role = $this->getById($id);
+
+        if (is_null($role)) {
+            abort(400,__('custom_message.role_plan_not_found'));
+        }
+
+        return view('/role/detail')->with('role', $role);
+    }
+
+    public function newForm() {
+        return view('/role/add_form');
+    }
+
+    public function updateForm($id) {
+        $role = $this->getById($id);
+
+        if (is_null($role)) {
+            abort(400,__('custom_message.role_plan_not_found'));
+        }
+
+        return view('/role/add_form')->with('role', $role);
     }
 
     public function getById($id)
@@ -29,14 +58,49 @@ class RoleService extends MyService
         return $this->roleRepo->getById($id);
     }
 
-    public function create($input)
+    public function create(Request $request)
     {
-        return $this->roleRepo->create($input);
+        //return $this->roleRepo->create($input);
+        $validator = $this->roleValidation->newValidate($request->all());
+
+        if ($validator->fails()) {
+            return redirect()->back()->withInput()->withErrors($validator->errors());
+        }
+
+        $input = $request->all();
+
+        try {
+
+            $role = $this->roleRepo->create($input);
+            return redirect()->intended('/system-admin/role/detail/' . $role->id)->with('saved', true);
+        } catch (\Exception $ex) {
+            abort(400,$ex->getMessage());
+        }
     }
 
-    public function update($id, $input)
+    public function update($id, $request)
     {
-        return $this->roleRepo->update($id, $input);
+        $role = $this->getById($id);
+
+        if (is_null($role)) {
+            abort(400,__('custom_message.role_plan_not_found'));
+        }
+
+        $validator = $this->roleValidation->updateValidate($request->all(), $id);
+
+        if ($validator->fails()) {
+            return redirect()->back()->withInput()->withErrors($validator->errors());
+        }
+
+        $input = request()->except(['_token', 'role']);
+
+        try {
+            $role = $this->roleRepo->update($id, $input);
+        } catch (\Exception $ex) {
+            abort(400,$ex->getMessage());
+        }
+
+        return redirect()->intended('/system-admin/role/detail/' . $id)->with('saved', true);
     }
 
     public function ruleCreateUpdate($request, $id = null)
@@ -60,7 +124,19 @@ class RoleService extends MyService
 
     public function delete($id)
     {
-        return $this->roleRepo->delete($id);
+        $role = $this->getById($id);
+
+        if (is_null($role)) {
+            abort(400,__('custom_message.role_plan_not_found'));
+        }
+
+        try {
+            $this->roleRepo->delete($id);
+        } catch (\Exception $ex) {
+            abort(400,$ex->getMessage());
+        }
+
+        return redirect()->intended('/system-admin/role')->with('deleted', true);
     }
 
 }
